+++ jdecode_test.go
@@ -1,21 +1,42 @@
 package junosdecode_test
 
 import (
+	"math/rand"
 	"testing"
 
-	. "github.com/jeremmfr/junosdecode"
+	"github.com/jeremmfr/junosdecode"
 )
 
-// TestDecode example password.
-func TestDecode(t *testing.T) {
-	junwordCoded := "$9$1HFIyKXxdsgJ-VH.Pfn6lKMXdsZUi5Qnikfz"
-	word := "testPassWord"
+const (
+	junWordCoded   = "$9$1HFIyKXxdsgJ-VH.Pfn6lKMXdsZUi5Qnikfz"
+	junWordDecoded = "testPassWord"
+)
 
-	passwordDecoded, err := Decode(junwordCoded)
+// TestDecode decode example password.
+func TestDecodePassword(t *testing.T) {
+	passwordDecoded, err := junosdecode.Decode(junWordCoded)
 	if err != nil {
 		t.Errorf("error on decode %v", err)
 	}
-	if passwordDecoded != word {
+	if passwordDecoded != junWordDecoded {
 		t.Errorf("decode password failed")
+	}
+}
+
+// nolint: gosec
+// TestDecodeBadEncoded try decode encrypted secret but with missing or extra character.
+func TestDecodeBadEncoded(t *testing.T) {
+	// remove last character
+	if passwordDecoded, err := junosdecode.Decode(junWordCoded[:len(junWordCoded)-1]); err == nil {
+		t.Errorf("missing character in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
+	}
+	// add last character
+	if passwordDecoded, err := junosdecode.Decode(junWordCoded + string(rune(rand.Intn(26)))); err == nil {
+		t.Errorf("extra character in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
+	}
+	// add two last characters
+	if passwordDecoded, err := junosdecode.Decode(
+		junWordCoded + string(rune(rand.Intn(26))) + string(rune(rand.Intn(26)))); err == nil {
+		t.Errorf("extra characters in junWordCoded not detected, passwordDecoded: %s", passwordDecoded)
 	}
 }